+++ UMFPACK/Source/umfpack_serialize_symbolic.c
@@ -23,7 +23,7 @@
 
 int UMFPACK_serialize_symbolic_size
 (
-    int64_t *blobsize,          // output: required size of blob
+    intptr_t *blobsize,         // output: required size of blob
     void *SymbolicHandle        // input: Symbolic object to serialize
 )
 {
@@ -45,7 +45,7 @@ int UMFPACK_serialize_symbolic_size
     }
 
     // blob header
-    (*blobsize) += sizeof (int64_t) + 10 * sizeof (int32_t) ;
+    (*blobsize) += sizeof (intptr_t) + 10 * sizeof (int32_t) ;
 
     // Symbolic header struct:
     (*blobsize) += sizeof (SymbolicType) ;
@@ -94,7 +94,7 @@ int UMFPACK_serialize_symbolic
 (
     int8_t *blob,           // output: serialized blob of size blobsize,
                             // allocated but unitialized on input.
-    int64_t blobsize,       // input: size of the blob
+    intptr_t blobsize,      // input: size of the blob
     void *SymbolicHandle    // input: Symbolic object to serialize
 )
 {
@@ -104,7 +104,7 @@ int UMFPACK_serialize_symbolic
     {
         return (UMFPACK_ERROR_argument_missing) ;
     }
-    int64_t required ;
+    intptr_t required ;
     int status = UMFPACK_serialize_symbolic_size (&required, SymbolicHandle) ;
     if (status != UMFPACK_OK)
     {
@@ -122,14 +122,14 @@ int UMFPACK_serialize_symbolic
 
     // write the blob header:
     int64_t offset = 0 ;
-    SERIALIZE (&required, int64_t, 1) ;         // required size of this blob
+    SERIALIZE (&required, intptr_t, 1) ;        // required size of this blob
     SERIALIZE_INT32 (SYMBOLIC_VALID) ;          // tag as a Symbolic object
     SERIALIZE_INT32 (UMFPACK_MAIN_VERSION) ;    // gaurd against version changes
     SERIALIZE_INT32 (UMFPACK_SUB_VERSION) ;
     SERIALIZE_INT32 (UMFPACK_SUBSUB_VERSION) ;
     SERIALIZE_INT32 (sizeof (SymbolicType)) ;   // size of Symbolic header
     SERIALIZE_INT32 (sizeof (Entry)) ;          // double or double complex
-    SERIALIZE_INT32 (sizeof (Int)) ;            // Int is int32_t or int64_t
+    SERIALIZE_INT32 (sizeof (Int)) ;            // Int is int32_t or SuiteSparse_long
     SERIALIZE_INT32 (sizeof (Unit)) ;
     SERIALIZE_INT32 (sizeof (double)) ;
     SERIALIZE_INT32 (sizeof (void *)) ;         // 32-bit vs 64-bit OS
